CASSSIDECAR-329: Refactored code based on review comments
