+++ adapters/adapters-base/src/main/java/org/apache/cassandra/sidecar/adapters/base/CassandraCompactionStatsOperations.java
@@ -16,58 +16,76 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.sidecar.service;
+package org.apache.cassandra.sidecar.adapters.base;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
+import org.apache.cassandra.sidecar.adapters.base.data.CompactionStatsMetrics;
+import org.apache.cassandra.sidecar.adapters.base.utils.DataTypeUtils;
 import org.apache.cassandra.sidecar.common.server.CompactionManagerOperations;
+import org.apache.cassandra.sidecar.common.server.CompactionStatsOperations;
 import org.apache.cassandra.sidecar.common.server.MetricsOperations;
 import org.apache.cassandra.sidecar.common.server.StorageOperations;
 import org.apache.cassandra.sidecar.common.server.data.ActiveCompactionEntryData;
 import org.apache.cassandra.sidecar.common.server.data.CompactionStatsData;
-import org.apache.cassandra.sidecar.common.server.data.CompactionStatsMetrics;
 import org.apache.cassandra.sidecar.common.server.data.CompletedCompactionsRateData;
-import org.apache.cassandra.sidecar.common.server.utils.DataTypeUtils;
 
-import static org.apache.cassandra.sidecar.adapters.base.CassandraMetricsOperations.COLUMNFAMILY;
-import static org.apache.cassandra.sidecar.adapters.base.CassandraMetricsOperations.COMPACTION_ID;
-import static org.apache.cassandra.sidecar.adapters.base.CassandraMetricsOperations.COMPLETED;
-import static org.apache.cassandra.sidecar.adapters.base.CassandraMetricsOperations.KEYSPACE;
-import static org.apache.cassandra.sidecar.adapters.base.CassandraMetricsOperations.SSTABLES;
-import static org.apache.cassandra.sidecar.adapters.base.CassandraMetricsOperations.TARGET_DIRECTORY;
-import static org.apache.cassandra.sidecar.adapters.base.CassandraMetricsOperations.TASK_TYPE;
-import static org.apache.cassandra.sidecar.adapters.base.CassandraMetricsOperations.TOTAL;
-import static org.apache.cassandra.sidecar.common.server.utils.DataTypeUtils.safeCast;
-import static org.apache.cassandra.sidecar.common.server.utils.DataTypeUtils.safeParseLong;
+import static org.apache.cassandra.sidecar.adapters.base.data.CompactionStatsMetrics.BYTES_COMPACTED;
+import static org.apache.cassandra.sidecar.adapters.base.data.CompactionStatsMetrics.COMPACTIONS_ABORTED;
+import static org.apache.cassandra.sidecar.adapters.base.data.CompactionStatsMetrics.COMPACTIONS_REDUCED;
+import static org.apache.cassandra.sidecar.adapters.base.data.CompactionStatsMetrics.PENDING_TASKS_BY_TABLE_NAME;
+import static org.apache.cassandra.sidecar.adapters.base.data.CompactionStatsMetrics.SSTABLES_DROPPED_FROM_COMPACTION;
+import static org.apache.cassandra.sidecar.adapters.base.data.CompactionStatsMetrics.TOTAL_COMPACTIONS_COMPLETED;
+import static org.apache.cassandra.sidecar.adapters.base.utils.DataTypeUtils.safeCast;
+import static org.apache.cassandra.sidecar.adapters.base.utils.DataTypeUtils.safeParseLong;
 
 /**
  * Service responsible for fetching compaction stats from various sources including
  * CassandraStorageOperations, CassandraMetricsOperations, and CompactionManagerOperations.
  */
-public class CompactionStatsService
+public class CassandraCompactionStatsOperations implements CompactionStatsOperations
 {
     // Default values
     // Default string value when actual value is unavailable
     public static final String DEFAULT_STRING_VALUE = "";
     // Default numeric value when actual value is unavailable
     public static final String DEFAULT_NUMBER_VALUE = "-1";
-    
+
+    // Constants for compaction info map keys
+    // Keyspace name being compacted
+    public static final String KEYSPACE = "keyspace";
+    // Column family (table) name being compacted
+    public static final String COLUMNFAMILY = "columnfamily";
+    // Number of bytes already processed in the compaction
+    public static final String COMPLETED = "completed";
+    // Total number of bytes to be processed in the compaction
+    public static final String TOTAL = "total";
+    // Type of compaction task (e.g., COMPACTION, VALIDATION, etc.)
+    public static final String TASK_TYPE = "taskType";
+    // Unique compaction identifier
+    public static final String COMPACTION_ID = "compactionId";
+    // Comma-separated list of SSTable names involved in the compaction
+    public static final String SSTABLES = "sstables";
+    // Directory where compaction output will be written
+    public static final String TARGET_DIRECTORY = "targetDirectory";
+
     private final StorageOperations storageOperations;
     private final MetricsOperations metricsOperations;
     private final CompactionManagerOperations compactionManagerOperations;
-    
-    public CompactionStatsService(StorageOperations storageOperations,
-                                  MetricsOperations metricsOperations,
-                                  CompactionManagerOperations compactionManagerOperations)
+
+    public CassandraCompactionStatsOperations(StorageOperations storageOperations,
+                                              MetricsOperations metricsOperations,
+                                              CompactionManagerOperations compactionManagerOperations)
     {
         this.storageOperations = storageOperations;
         this.metricsOperations = metricsOperations;
         this.compactionManagerOperations = compactionManagerOperations;
     }
-    
+
+    @Override
     public CompactionStatsData compactionStats()
     {
         // Get concurrent compactors from StorageOperations
@@ -76,20 +94,15 @@ public class CompactionStatsService
         // Get pending tasks grouped by keyspace and table
         Map<String, Map<String, Integer>> pendingTasks = getPendingCompactionTasksByTable();
         long totalPendingTasks = pendingTasks.values().stream()
-                                                   .mapToLong(tableMap -> tableMap.values().stream().mapToInt(Integer::intValue).sum())
-                                                   .sum();
+                                             .mapToLong(tableMap -> tableMap.values().stream().mapToInt(Integer::intValue).sum())
+                                             .sum();
 
         // Get compaction metrics from JMX counters and meters
-        long completedCompactionsCount = DataTypeUtils.getValueAsLong(
-                metricsOperations.getCompactionMetric(CompactionStatsMetrics.TOTAL_COMPACTIONS_COMPLETED));
-        long dataCompactedBytes = DataTypeUtils.getValueAsLong(
-                metricsOperations.getCompactionMetric(CompactionStatsMetrics.BYTES_COMPACTED));
-        long abortedCompactionsCount = DataTypeUtils.getValueAsLong(
-                metricsOperations.getCompactionMetric(CompactionStatsMetrics.COMPACTIONS_ABORTED));
-        long reducedCompactionsCount = DataTypeUtils.getValueAsLong(
-                metricsOperations.getCompactionMetric(CompactionStatsMetrics.COMPACTIONS_REDUCED));
-        long sstablesDroppedFromCompactionCount = DataTypeUtils.getValueAsLong(
-                metricsOperations.getCompactionMetric(CompactionStatsMetrics.SSTABLES_DROPPED_FROM_COMPACTION));
+        long completedCompactionsCount = getCompactionMetricLong(TOTAL_COMPACTIONS_COMPLETED);
+        long dataCompactedBytes = getCompactionMetricLong(BYTES_COMPACTED);
+        long abortedCompactionsCount = getCompactionMetricLong(COMPACTIONS_ABORTED);
+        long reducedCompactionsCount = getCompactionMetricLong(COMPACTIONS_REDUCED);
+        long sstablesDroppedFromCompactionCount = getCompactionMetricLong(SSTABLES_DROPPED_FROM_COMPACTION);
 
         // Get completed compactions rate with proper time conversions
         CompletedCompactionsRateData completedCompactionsRate = metricsOperations.getCompletedCompactionsRate();
@@ -102,25 +115,30 @@ public class CompactionStatsService
         long activeCompactionsRemainingTime = calculateRemainingTimeSeconds(activeCompactions);
 
         return CompactionStatsData.builder()
-            .concurrentCompactors(concurrentCompactors)
-            .pendingTasks(pendingTasks)
-            .totalPendingTasks(totalPendingTasks)
-            .completedCompactions(completedCompactionsCount)
-            .dataCompacted(dataCompactedBytes)
-            .abortedCompactions(abortedCompactionsCount)
-            .reducedCompactions(reducedCompactionsCount)
-            .sstablesDroppedFromCompaction(sstablesDroppedFromCompactionCount)
-            .completedCompactionsRate(completedCompactionsRate)
-            .activeCompactions(activeCompactions)
-            .activeCompactionsCount(activeCompactionsCount)
-            .activeCompactionsRemainingTime(activeCompactionsRemainingTime)
-            .build();
+                                  .concurrentCompactors(concurrentCompactors)
+                                  .pendingTasks(pendingTasks)
+                                  .totalPendingTasks(totalPendingTasks)
+                                  .completedCompactions(completedCompactionsCount)
+                                  .dataCompacted(dataCompactedBytes)
+                                  .abortedCompactions(abortedCompactionsCount)
+                                  .reducedCompactions(reducedCompactionsCount)
+                                  .sstablesDroppedFromCompaction(sstablesDroppedFromCompactionCount)
+                                  .completedCompactionsRate(completedCompactionsRate)
+                                  .activeCompactions(activeCompactions)
+                                  .activeCompactionsCount(activeCompactionsCount)
+                                  .activeCompactionsRemainingTime(activeCompactionsRemainingTime)
+                                  .build();
+    }
+
+    private long getCompactionMetricLong(CompactionStatsMetrics metric)
+    {
+        return DataTypeUtils.getValueAsLong(metricsOperations.getCompactionMetric(metric.metricName(), metric.type));
     }
 
     private Map<String, Map<String, Integer>> getPendingCompactionTasksByTable()
     {
         // Get pending tasks by table name from Gauge metric
-        Object value = metricsOperations.getCompactionMetric(CompactionStatsMetrics.PENDING_TASKS_BY_TABLE_NAME);
+        Object value = metricsOperations.getCompactionMetric(PENDING_TASKS_BY_TABLE_NAME.metricName(), PENDING_TASKS_BY_TABLE_NAME.type);
         return parsePendingTasksMap(value);
     }
 
@@ -175,16 +193,16 @@ public class CompactionStatsService
             String targetDirectory = compactionInfo.getOrDefault(TARGET_DIRECTORY, DEFAULT_STRING_VALUE);
 
             return ActiveCompactionEntryData.builder()
-                .id(id)
-                .keyspace(keyspace)
-                .table(table)
-                .taskType(taskType)
-                .completedBytes(completedBytes)
-                .totalBytes(totalBytes)
-                .percentCompleted(percentCompleted)
-                .sstables(ssTables)
-                .targetDirectory(targetDirectory)
-                .build();
+                                            .id(id)
+                                            .keyspace(keyspace)
+                                            .table(table)
+                                            .taskType(taskType)
+                                            .completedBytes(completedBytes)
+                                            .totalBytes(totalBytes)
+                                            .percentCompleted(percentCompleted)
+                                            .sstables(ssTables)
+                                            .targetDirectory(targetDirectory)
+                                            .build();
         }).collect(Collectors.toList());
     }
 
@@ -197,9 +215,9 @@ public class CompactionStatsService
 
         // Calculate total remaining bytes across all active compactions
         long totalRemainingBytes = activeCompactions.stream()
-                .filter(compaction -> compaction.totalBytes >= 0 && compaction.completedBytes >= 0)
-                .mapToLong(compaction -> Math.max(0, compaction.totalBytes - compaction.completedBytes))
-                .sum();
+                                                    .filter(compaction -> compaction.totalBytes >= 0 && compaction.completedBytes >= 0)
+                                                    .mapToLong(compaction -> Math.max(0, compaction.totalBytes - compaction.completedBytes))
+                                                    .sum();
 
         long throughputBytesPerSec = storageOperations.getCompactionThroughputBytesPerSec();
         if (totalRemainingBytes < 0 || throughputBytesPerSec <= 0)
