+++ adapters/adapters-base/src/main/java/org/apache/cassandra/sidecar/adapters/base/CassandraStorageOperations.java
@@ -30,6 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.apache.cassandra.sidecar.adapters.base.jmx.StorageJmxOperations;
+import org.apache.cassandra.sidecar.adapters.base.utils.DataTypeUtils;
 import org.apache.cassandra.sidecar.common.response.RingResponse;
 import org.apache.cassandra.sidecar.common.response.TokenRangeReplicasResponse;
 import org.apache.cassandra.sidecar.common.server.JmxClient;
@@ -308,10 +309,18 @@ public class CassandraStorageOperations implements StorageOperations
     /**
      * {@inheritDoc}
      */
+    public long getCompactionThroughputMbPerSec()
+    {
+        return jmxClient.proxy(StorageJmxOperations.class, STORAGE_SERVICE_OBJ_NAME)
+                        .getCompactionThroughputMbPerSec();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public long getCompactionThroughputBytesPerSec()
     {
-        LOGGER.warn("getCompactionThroughputBytesPerSec is not supported in Cassandra 4.0");
-        return 0;
+        return DataTypeUtils.megabytesToBytes(getCompactionThroughputMbPerSec());
     }
 }
