+++ server-common/src/main/java/org/apache/cassandra/sidecar/common/server/ICassandraAdapter.java
@@ -118,4 +118,10 @@ public interface ICassandraAdapter
      * @throws CassandraUnavailableException when Cassandra is not available
      */
     @NotNull CompactionManagerOperations compactionManagerOperations() throws CassandraUnavailableException;
+
+    /**
+     * @return the {@link CompactionStatsOperations} implementation for the Cassandra cluster
+     * @throws CassandraUnavailableException when Cassandra is not available
+     */
+    @NotNull CompactionStatsOperations compactionStatsOperations() throws CassandraUnavailableException;
 }
