+++ server/src/main/java/org/apache/cassandra/sidecar/cluster/CassandraAdapterDelegate.java
@@ -49,6 +49,7 @@ import org.apache.cassandra.sidecar.common.response.NodeSettings;
 import org.apache.cassandra.sidecar.common.server.CQLSessionProvider;
 import org.apache.cassandra.sidecar.common.server.ClusterMembershipOperations;
 import org.apache.cassandra.sidecar.common.server.CompactionManagerOperations;
+import org.apache.cassandra.sidecar.common.server.CompactionStatsOperations;
 import org.apache.cassandra.sidecar.common.server.ICassandraAdapter;
 import org.apache.cassandra.sidecar.common.server.JmxClient;
 import org.apache.cassandra.sidecar.common.server.MetricsOperations;
@@ -433,6 +434,13 @@ public class CassandraAdapterDelegate implements ICassandraAdapter, Host.StateLi
     public CompactionManagerOperations compactionManagerOperations() throws CassandraUnavailableException
     {
         return fromAdapter(ICassandraAdapter::compactionManagerOperations);
+    }
+
+    @Override
+    @NotNull
+    public CompactionStatsOperations compactionStatsOperations() throws CassandraUnavailableException
+    {
+        return fromAdapter(ICassandraAdapter::compactionStatsOperations);
     }
 
     @Override
